Comment cleanup, renames, removed old Pets module, updated Grimoire: Felguard module, Changelog entry
