+++ src/parser/warlock/demonology/modules/talents/GrimoireFelguard.js
@@ -4,18 +4,18 @@ import Analyzer from 'parser/core/Analyzer';
 import AbilityTracker from 'parser/shared/modules/AbilityTracker';
 
 import SPELLS from 'common/SPELLS';
-import PETS from 'common/PETS';
-import { formatNumber, formatPercentage } from 'common/format';
+import { formatNumber, formatThousands } from 'common/format';
 import SpellIcon from 'common/SpellIcon';
 
 import StatisticBox, { STATISTIC_ORDER } from 'interface/others/StatisticBox';
 
-import OldPets from '../pets/OldPets';
+import Pets from '../pets/Pets';
+import PET_INFO from '../pets/PET_INFO';
 
 class GrimoireFelguard extends Analyzer {
   static dependencies = {
     abilityTracker: AbilityTracker,
-    oldPets: OldPets,
+    pets: Pets,
   };
 
   constructor(...args) {
@@ -24,14 +24,13 @@ class GrimoireFelguard extends Analyzer {
   }
 
   statistic() {
-    // TODO: Probably fix felguard GUID in PETS
-    const damage = this.oldPets.getTotalPetDamage(PETS.GRIMOIRE_FELGUARD.id);
+    const damage = this.pets.getPetDamage(PET_INFO.GRIMOIRE_FELGUARD.guid);
     return (
       <StatisticBox
         icon={<SpellIcon id={SPELLS.GRIMOIRE_FELGUARD_TALENT.id} />}
         value={`${formatNumber(damage / this.owner.fightDuration * 1000)} DPS`}
         label="Grimoire: Felguard damage"
-        tooltip={`Your Grimoire: Felguard did ${formatNumber(damage)} damage (${formatPercentage(this.owner.getPercentageOfTotalDamageDone(damage))} %).`}
+        tooltip={`Your Grimoire: Felguard did ${formatThousands(damage)} damage (${this.owner.formatItemDamageDone(damage)}).`}
       />
     );
   }
